+++ src/shared/ui/formField/rhf/RHFTextField.tsx
@@ -1,6 +1,6 @@
 import { Control, Controller, FieldError, FieldPath, FieldValues } from 'react-hook-form'
 import { FormField } from '../FormField'
-import { hasError, isEmptyValue } from '@/features/add-book/utils'
+import { hasError, isEmptyValue } from '@/shared/utils'
 import { TextField } from '../../textField'
 
 interface RHFTextFieldProps<TFieldValues extends FieldValues> {
